requirements.txt
asgiref==3.2.7
astroid==2.4.1
cffi==1.14.0
colorama==0.4.3
cryptography==2.9.2
dj-database-url==0.5.0
Django==3.0.6
django-cors-headers==3.2.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.11.0
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
psycopg2==2.8.5
pycparser==2.20
pylint==2.5.2
pytz==2020.1
six==1.14.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.0.1
wrapt==1.12.1
